+++ service/leshuazf/client.go
@@ -1,54 +1,39 @@
 package leshuazf
 
 import (
-	"github.com/dtapps/go-library/utils/dorm"
 	"github.com/dtapps/go-library/utils/golog"
 	"github.com/dtapps/go-library/utils/gorequest"
-	"gorm.io/gorm"
 )
 
 type ConfigClient struct {
-	AgentId      string // 服务商编号，由乐刷分配的接入方唯一标识，明文传输。
-	Environment  string //  环境
-	KeyAgent     string
-	MongoDb      *dorm.MongoClient // 日志数据库
-	PgsqlDb      *gorm.DB          // 日志数据库
-	DatabaseName string            // 库名
+	AgentId     string // 服务商编号，由乐刷分配的接入方唯一标识，明文传输。
+	Environment string // 环境
+	KeyAgent    string
 }
 
 // Client 乐刷
 type Client struct {
-	client *gorequest.App   // 请求客户端
-	log    *golog.ApiClient // 日志服务
-	config *ConfigClient    // 配置
+	client *gorequest.App // 请求服务
+	config struct {
+		agentId     string // 服务商编号，由乐刷分配的接入方唯一标识，明文传输。
+		environment string // 环境
+		keyAgent    string
+	}
+	log struct {
+		status bool             // 状态
+		client *golog.ApiClient // 日志服务
+	}
 }
 
 func NewClient(config *ConfigClient) (*Client, error) {
 
-	var err error
-	c := &Client{config: config}
+	c := &Client{}
+
+	c.config.agentId = config.AgentId
+	c.config.environment = config.Environment
+	c.config.keyAgent = config.KeyAgent
 
 	c.client = gorequest.NewHttp()
-
-	if c.config.PgsqlDb != nil {
-		c.log, err = golog.NewApiClient(
-			golog.WithGormClient(c.config.PgsqlDb),
-			golog.WithTableName(logTable),
-		)
-		if err != nil {
-			return nil, err
-		}
-	}
-	if c.config.MongoDb != nil {
-		c.log, err = golog.NewApiClient(
-			golog.WithMongoClient(c.config.MongoDb),
-			golog.WithDatabaseName(c.config.DatabaseName),
-			golog.WithCollectionName(logTable),
-		)
-		if err != nil {
-			return nil, err
-		}
-	}
 
 	return c, nil
 }