+++ service/leshuazf/request.go
@@ -1,12 +1,14 @@
 package leshuazf
 
 import (
+	"context"
+	go_library "github.com/dtapps/go-library"
 	"github.com/dtapps/go-library/utils/gorandom"
 	"github.com/dtapps/go-library/utils/gorequest"
 	"github.com/dtapps/go-library/utils/gotime"
 )
 
-func (c *Client) request(url string, params map[string]interface{}, method string) (gorequest.Response, error) {
+func (c *Client) request(ctx context.Context, url string, params map[string]interface{}, method string) (gorequest.Response, error) {
 
 	// 环境
 	if c.GetEnvironment() == "test" {
@@ -37,17 +39,14 @@ func (c *Client) request(url string, params map[string]interface{}, method strin
 	client.SetParams(params)
 
 	// 发起请求
-	request, err := client.Request()
+	request, err := client.Request(ctx)
 	if err != nil {
 		return gorequest.Response{}, err
 	}
 
 	// 日志
-	if c.config.PgsqlDb != nil {
-		go c.log.GormMiddleware(request)
-	}
-	if c.config.MongoDb != nil {
-		go c.log.MongoMiddleware(request)
+	if c.log.status {
+		go c.log.client.Middleware(ctx, request, go_library.Version())
 	}
 
 	return request, err