+++ service/pconline/request.go
@@ -1,8 +1,12 @@
 package pconline
 
-import "github.com/dtapps/go-library/utils/gorequest"
+import (
+	"context"
+	go_library "github.com/dtapps/go-library"
+	"github.com/dtapps/go-library/utils/gorequest"
+)
 
-func (c *Client) request(url string) (gorequest.Response, error) {
+func (c *Client) request(ctx context.Context, url string, params map[string]interface{}) (gorequest.Response, error) {
 
 	// 创建请求
 	client := c.client
@@ -10,18 +14,18 @@ func (c *Client) request(url string) (gorequest.Response, error) {
 	// 设置请求地址
 	client.SetUri(url)
 
+	// 设置参数
+	client.SetParams(params)
+
 	// 发起请求
-	request, err := client.Get()
+	request, err := client.Get(ctx)
 	if err != nil {
 		return gorequest.Response{}, err
 	}
 
 	// 日志
-	if c.config.PgsqlDb != nil {
-		go c.log.GormMiddleware(request)
-	}
-	if c.config.MongoDb != nil {
-		go c.log.MongoMiddleware(request)
+	if c.log.status {
+		go c.log.client.Middleware(ctx, request, go_library.Version())
 	}
 
 	return request, err